[FEAT] 컨트롤러 함수 생성 및 mypage 폴더 구조 수정
